+++ src/jmx.go
@@ -19,7 +19,7 @@ type argumentList struct {
 	JmxPass         string `default:"admin" help:"The password for the JMX connection"`
 	CollectionFiles string `default:"" help:"A comma separated list of full paths to metrics configuration files"`
 	Timeout         int    `default:"10000" help:"Timeout for JMX queries"`
-	MetricLimit     int    `default:"200" help:"Number of metrics that can be collected per entity. If this limit is exceeded some metrics may be lost. A limit of 0 implies not limit."`
+	MetricLimit     int    `default:"200" help:"Number of metrics that can be collected per entity. If this limit is exceeded the entity will not be reported. A limit of 0 implies no limit."`
 }
 
 const (
@@ -90,7 +90,7 @@ func main() {
 
 	jmxCloseFunc()
 
-	checkMetricLimit(jmxIntegration.Entities)
+	jmxIntegration.Entities = checkMetricLimit(jmxIntegration.Entities)
 
 	if err := jmxIntegration.Publish(); err != nil {
 		os.Exit(1)
@@ -99,7 +99,9 @@ func main() {
 
 // checkMetricLimit looks through all of the metric sets for every entity and aggregates the number
 // of metrics. If that total is greate than args.MetricLimit a warning is logged
-func checkMetricLimit(entities []*integration.Entity) {
+func checkMetricLimit(entities []*integration.Entity) []*integration.Entity {
+	validEntities := make([]*integration.Entity, 0, len(entities))
+
 	for _, entity := range entities {
 		metricCount := 0
 		for _, metricSet := range entity.Metrics {
@@ -107,7 +109,12 @@ func checkMetricLimit(entities []*integration.Entity) {
 		}
 
 		if args.MetricLimit != 0 && metricCount > args.MetricLimit {
-			log.Warn("Domain '%s' has %d metrics, the current limit is %d. Some metrics will be truncated.", entity.Metadata.Name, metricCount, args.MetricLimit)
+			log.Warn("Domain '%s' has %d metrics, the current limit is %d. This Domain will not be reported", entity.Metadata.Name, metricCount, args.MetricLimit)
+			continue
 		}
+
+		validEntities = append(validEntities, entity)
 	}
+
+	return validEntities
 }