+++ src/jmx_test.go
@@ -7,15 +7,9 @@ import (
 	"github.com/newrelic/infra-integrations-sdk/integration"
 )
 
-func Test_checkMetricList_NoPanic(t *testing.T) {
+func Test_checkMetricList(t *testing.T) {
 	// Ensure we hit the
-	args.MetricLimit = 1
-
-	defer func() {
-		if r := recover(); r != nil {
-			t.Errorf("Unexpected panic: %s", r)
-		}
-	}()
+	args.MetricLimit = 2
 
 	i, err := integration.New("kafka", "1.0.0")
 	if err != nil {
@@ -23,14 +17,33 @@ func Test_checkMetricList_NoPanic(t *testing.T) {
 		t.FailNow()
 	}
 
-	e, err := i.Entity("test", "domain")
+	e1, err := i.Entity("test_1", "domain")
 	if err != nil {
 		t.Errorf("Unexpected error: %s", err.Error())
 		t.FailNow()
 	}
 
-	set := e.NewMetricSet("testSample")
-	set.SetMetric("new_metric", 4, metric.GAUGE)
+	set1 := e1.NewMetricSet("testSample")
+	set1.SetMetric("new_metric", 4, metric.GAUGE)
 
-	checkMetricLimit(i.Entities)
+	e2, err := i.Entity("test_2", "domain")
+	if err != nil {
+		t.Errorf("Unexpected error: %s", err.Error())
+		t.FailNow()
+	}
+
+	set2 := e2.NewMetricSet("testSample")
+	set2.SetMetric("new_metric", 4, metric.GAUGE)
+	set2.SetMetric("other_metric", 5, metric.GAUGE)
+
+	out := checkMetricLimit(i.Entities)
+
+	if length := len(out); length != 1 {
+		t.Errorf("Expected 1 entity got %d", length)
+		t.FailNow()
+	}
+
+	if out[0] != e1 {
+		t.Errorf("Expected entity '%+v' got '%+v'", e1, out[0])
+	}
 }